fix : 컴포넌트 변경(데이터, 타입 분리)
